Real-Time Streaming Data Pipeline with Kafka, PySpark, and MinIO/requirements.txt
kafka-python
pyspark
boto3
pandas
fastparquet
pyarrow